Create tableau_03_ransom.cpp
Not sure whether there is any missing points here. Haven't run the code yet.
In the end, pointed out that we can use one unordered_map to increment and decrement... 
I think this is a problem in cracking the code, but I didn't remember the optimized answer at the time of this test.